#44 Restructure
